+++ apps/mobile/src/stores/preferences.ts
@@ -51,6 +51,7 @@ export type PreferencesPayload = {
   largeThumbnails: boolean
   linkBrowser: boolean
   mediaOnRight: boolean
+  minimizeTabBar: boolean
   oldReddit: boolean
   pictureInPicture: boolean
   privateScreenshots: boolean
@@ -72,7 +73,6 @@ export type PreferencesPayload = {
   sortUserPosts: UserFeedSort
   stickyDrawer: boolean
   systemScaling: boolean
-  tabBarMinimizeBehavior: 'never' | 'automatic' | 'onScrollDown' | 'onScrollUp'
   theme: Theme
   themeOled: boolean
   themeTint: boolean
@@ -119,6 +119,7 @@ export const preferencesStore = create<State>()(
       largeThumbnails: false,
       linkBrowser: true,
       mediaOnRight: true,
+      minimizeTabBar: false,
       oldReddit: false,
       pictureInPicture: false,
       privateScreenshots: true,
@@ -140,7 +141,6 @@ export const preferencesStore = create<State>()(
       sortUserPosts: 'new',
       stickyDrawer: true,
       systemScaling: false,
-      tabBarMinimizeBehavior: 'never',
       theme: 'acorn',
       themeOled: false,
       themeTint: true,